integrate kustomize with spot interrupt test

+++ test/spot-termination-test/run-spot-termination-test.sh
@@ -17,7 +17,12 @@ NODE_TERMINATION_HANDLER_DOCKER_IMG=""
 DEFAULT_EC2_METADATA_DOCKER_IMG="ec2-meta-data-proxy:customtest"
 EC2_METADATA_DOCKER_IMG=""
 
-K8_DEPLOYMENT_SPEC="spot-termination-test.yaml"
+TMP_DIR=$SCRIPTPATH/tmp-$TEST_ID
+
+KUSTOMIZATION_FILE="$TMP_DIR/kustomization.yaml"
+NTH_OVERLAY_FILE="nth-image-overlay.yaml"
+METADATA_OVERLAY_FILE="ec2-metadata-image-overlay.yaml"
+REGULAR_POD_OVERLAY_FILE="ec2-metadata-regular-pod-image-overlay.yaml"
 
 K8_1_16="kindest/node:v1.16.2@sha256:5bbdfa140633b135672ff0e1eb1a1b37afcab36216103c0b3d97337c62c5e2a1"
 K8_1_15="kindest/node:v1.15.3@sha256:27e388752544890482a86b90d8ac50fcfa63a2e8656a96ec5337b902ec8e5157"
@@ -83,14 +88,15 @@ done
 CLUSTER_NAME="$CLUSTER_NAME_BASE"-"$TEST_ID"
 echo "🐳 Using Kubernetes $K8_VERSION"
 
-TMP_DIR=$SCRIPTPATH/tmp-$TEST_ID
 ## Append to the end of PATH so that the user can override the executables if they want
 export PATH=$PATH:$TMP_DIR
 mkdir -p $TMP_DIR
 
 function clean_up {
-    [[ "$PRESERVE" == false ]] && kind delete cluster --name $CLUSTER_NAME
-    rm -rf $TMP_DIR
+    if [[ "$PRESERVE" == false ]]; then
+        kind delete cluster --name $CLUSTER_NAME
+        rm -rf $TMP_DIR
+    fi 
 }
 
 function exit_and_fail {
@@ -133,13 +139,35 @@ kind create cluster --name "$CLUSTER_NAME" --image $K8_VERSION --config "$SCRIPT
 export KUBECONFIG="$(kind get kubeconfig-path --name=$CLUSTER_NAME)"
 echo "👍 Created k8s cluster using \"kind\" and added kube config to KUBECONFIG"
 
+cat <<-EOF > $KUSTOMIZATION_FILE
+apiVersion: kustomize.config.k8s.io/v1beta1
+kind: Kustomization
+
+bases:
+- ../../../config/overlays/test
+
+patchesStrategicMerge:
+EOF
+
 if [ -z "$NODE_TERMINATION_HANDLER_DOCKER_IMG" ]; then 
     echo "🥑 Building the node-termination-handler docker image"
     docker build $DOCKER_ARGS -t $DEFAULT_NODE_TERMINATION_HANDLER_DOCKER_IMG --no-cache --force-rm "$SCRIPTPATH/../../." 
     NODE_TERMINATION_HANDLER_DOCKER_IMG="$DEFAULT_NODE_TERMINATION_HANDLER_DOCKER_IMG"
     echo "👍 Built the node-termination-handler docker image"
 else 
-    sed -i "s+$DEFAULT_NODE_TERMINATION_HANDLER_DOCKER_IMG+$NODE_TERMINATION_HANDLER_DOCKER_IMG+" "$SCRIPTPATH/$K8_DEPLOYMENT_SPEC"
+    cat <<-EOF > $TMP_DIR/$NTH_OVERLAY_FILE
+apiVersion: apps/v1
+kind: DaemonSet
+metadata:
+  name: node-termination-handler
+spec:
+  template:
+    spec:
+      containers:
+      - name: node-termination-handler
+        image: $NODE_TERMINATION_HANDLER_DOCKER_IMG
+EOF
+    echo "- $NTH_OVERLAY_FILE" >> $KUSTOMIZATION_FILE
     echo "🥑 Skipping building the node-termination-handler docker image, since one was specified ($NODE_TERMINATION_HANDLER_DOCKER_IMG)"
 fi
 
@@ -149,9 +177,36 @@ if [ -z "$EC2_METADATA_DOCKER_IMG" ]; then
     EC2_METADATA_DOCKER_IMG="$DEFAULT_EC2_METADATA_DOCKER_IMG"
     echo "👍 Built the ec2-meta-data-proxy docker image"
 else 
-    sed -i "s+$DEFAULT_EC2_METADATA_DOCKER_IMG+$EC2_METADATA_DOCKER_IMG+" "$SCRIPTPATH/$K8_DEPLOYMENT_SPEC"
+    cat <<-EOF > $TMP_DIR/$METADATA_OVERLAY_FILE
+apiVersion: apps/v1
+kind: DaemonSet
+metadata:
+  name: node-termination-handler
+spec:
+  template:
+    spec:
+      containers:
+      - name: ec2-metadata-proxy
+        image: $EC2_METADATA_DOCKER_IMG
+EOF
+    echo "- $METADATA_OVERLAY_FILE" >> $KUSTOMIZATION_FILE
+        cat <<-EOF > $TMP_DIR/$REGULAR_POD_OVERLAY_FILE
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: regular-pod-test
+  namespace: default
+spec:
+  template:
+    spec:
+      containers:
+      - name: ec2-meta-data-proxy
+        image: $EC2_METADATA_DOCKER_IMG
+EOF
+    echo "- $REGULAR_POD_OVERLAY_FILE" >> $KUSTOMIZATION_FILE
     echo "🥑 Skipping building the ec2-meta-data-proxy docker image, since one was specified ($EC2_METADATA_DOCKER_IMG)"
 fi
+
 echo "🥑 Tagging worker nodes to execute integ test"
 kubectl label nodes $CLUSTER_NAME-worker lifecycle=Ec2Spot
 kubectl label nodes $CLUSTER_NAME-worker app=spot-termination-test
@@ -163,7 +218,7 @@ kind load docker-image --name $CLUSTER_NAME --nodes=$CLUSTER_NAME-worker,$CLUSTE
 echo "👍 Loaded both images into the cluster"
 
 echo "🥑 Applying the test overlay kustomize config to k8s using kubectl"
-kubectl apply -k "$SCRIPTPATH/../../config/overlays/test"
+kubectl apply -k "$TMP_DIR"
 
 for i in `seq 1 $TAINT_CHECK_CYCLES`; do
     if kubectl get events | grep regular-pod-test | grep Started; then
